+++ modules/imaging_qc/jsx/imagingQCIndex.js
@@ -3,7 +3,14 @@ import Loader from 'Loader';
 import FilterableDataTable from 'FilterableDataTable';
 import PropTypes from 'prop-types';
 
+/**
+ * Imaging Quality Control React Component
+ */
 class ImagingQCIndex extends Component {
+  /**
+   * @constructor
+   * @param {object} props - React Component properties
+   */
   constructor(props) {
     super(props);
     this.state = {
@@ -23,21 +30,25 @@ class ImagingQCIndex extends Component {
    *
    * @return {*} a formated table cell for a given column
    */
-
   formatColumn(column, cell, row) {
       let result = <td>{cell}</td>;
       switch (column) {
       case 'Scan Done in MRI PF':
         if (cell == 'Yes') {
-          let mpfURL = loris.BaseURL + '/instruments/mri_parameter_form/?commentID=' +
-              row.CommentID + '&sessionID=' + row['Session ID'] +
-              '&candID=' + row.DCCID;
+          let mpfURL = loris.BaseURL
+                       + '/instruments/mri_parameter_form/?commentID='
+                       + row.CommentID
+                       + '&sessionID='
+                       + row['Session ID']
+                       + '&candID='
+                       + row.DCCID;
           result = <td><a href={mpfURL}>{cell}</a></td>;
         }
       case 'Scan Location':
         if (cell == 'In Imaging Browser') {
-          let imgURL = loris.BaseURL + '/imaging_browser/viewSession/?sessionID=' +
-              row['Session ID'];
+          let imgURL = loris.BaseURL
+                       + '/imaging_browser/viewSession/?sessionID='
+                       + row['Session ID'];
           result = <td><a href={imgURL}>{cell}</a></td>;
         }
       case 'Tarchive':
@@ -50,11 +61,22 @@ class ImagingQCIndex extends Component {
       return result;
   }
 
+  /**
+   * Called by React when the component has been rendered on the page.
+   */
   componentDidMount() {
     this.fetchData(this.props.ImgDataURL, 'ImgData')
         .then(() => this.setState({isLoadedImg: true}));
   }
 
+  /**
+   * Retrive data from the provided URL and save it in state
+   *
+   * @param {string} url
+   * @param {object} state - The React state object
+   *
+   * @return {object}
+   */
   fetchData(url, state) {
     return fetch(url, {credentials: 'same-origin'})
         .then((resp) => resp.json())
@@ -64,6 +86,11 @@ class ImagingQCIndex extends Component {
       });
   }
 
+  /**
+   * Renders the React component.
+   *
+   * @return {JSX} - React markup for the component
+   */
   render() {
     if (!this.state.isLoadedImg) {
       return <Loader/>;