+++ modules/media/jsx/uploadForm.js
@@ -2,6 +2,7 @@ import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import ProgressBar from 'ProgressBar';
 import Loader from 'jsx/Loader';
+import swal from 'sweetalert2';
 
 /**
  * Media Upload Form
@@ -86,8 +87,10 @@ class MediaUploadForm extends Component {
     const visits = this.state.formData.pscid ?
       this.state.Data.sessionData[this.state.formData.pscid].visits :
       {};
-    const instruments = this.state.formData.pscid && this.state.formData.visitLabel ?
-      this.state.Data.sessionData[this.state.formData.pscid].instruments[this.state.formData.visitLabel] :
+    const instruments = this.state.formData.pscid
+                        && this.state.formData.visitLabel ?
+      this.state.Data.sessionData[this.state.formData.pscid]
+        .instruments[this.state.formData.visitLabel] :
       {};
     return (
       <div className='row'>
@@ -209,7 +212,9 @@ class MediaUploadForm extends Component {
 
     let formData = this.state.formData;
     let formRefs = this.refs;
-    let mediaFiles = this.state.Data.mediaFiles ? this.state.Data.mediaFiles : [];
+    let mediaFiles = this.state.Data.mediaFiles ?
+      this.state.Data.mediaFiles :
+      [];
 
     // Validate the form
     if (!this.isValidForm(formRefs, formData)) {
@@ -218,12 +223,14 @@ class MediaUploadForm extends Component {
 
     // Validate uploaded file name
     let instrument = formData.instrument ? formData.instrument : null;
-    let fileName = formData.file ? formData.file.name.replace(/\s+/g, '_') : null;
+    let fileName = formData.file ?
+      formData.file.name.replace(/\s+/g, '_') :
+      null;
     let requiredFileName = this.getValidFileName(
       formData.pscid, formData.visitLabel, instrument
     );
     if (!this.isValidFileName(requiredFileName, fileName)) {
-      swal(
+      swal.fire(
         'Invalid file name!',
         'File name should begin with: ' + requiredFileName,
         'error'
@@ -234,9 +241,10 @@ class MediaUploadForm extends Component {
     // Check for duplicate file names
     let isDuplicate = mediaFiles.indexOf(fileName);
     if (isDuplicate >= 0) {
-      swal({
+      swal.fire({
         title: 'Are you sure?',
-        text: 'A file with this name already exists!\n Would you like to override existing file?',
+        text: 'A file with this name already exists!\n '
+              + 'Would you like to override existing file?',
         type: 'warning',
         showCancelButton: true,
         confirmButtonText: 'Yes, I am sure!',
@@ -245,7 +253,7 @@ class MediaUploadForm extends Component {
         if (isConfirm) {
           this.uploadFile();
         } else {
-          swal('Cancelled', 'Your imaginary file is safe :)', 'error');
+          swal.fire('Cancelled', 'Your imaginary file is safe :)', 'error');
         }
       }.bind(this));
     } else {
@@ -297,7 +305,7 @@ class MediaUploadForm extends Component {
           formData: {}, // reset form data after successful file upload
           uploadProgress: -1,
         });
-        swal('Upload Successful!', '', 'success');
+        swal.fire('Upload Successful!', '', 'success');
       }.bind(this),
       error: function(err) {
         console.error(err);
@@ -306,7 +314,7 @@ class MediaUploadForm extends Component {
           errorMessage: msg,
           uploadProgress: -1,
         });
-        swal(msg, '', 'error');
+        swal.fire(msg, '', 'error');
       }.bind(this),
     });
   }