+++ modules/publication/jsx/viewProjectIndex.js
@@ -8,8 +8,12 @@ $(function() {
       <div className="row">
         <div className="col-md-9 col-lg-7">
           <ViewProject
-            DataURL={`${loris.BaseURL}/publication/ajax/getData.php?action=getProjectData&id=${args.id}`}
-            action={`${loris.BaseURL}/publication/ajax/FileUpload.php?action=editProject&id=${args.id}`}
+            DataURL={loris.BaseURL
+                     + '/publication/ajax/getData.php?action=getProjectData&id='
+                     + args.id}
+            action={loris.BaseURL
+                    + '/publication/ajax/FileUpload.php?action=editProject&id='
+                    + args.id}
           />
         </div>
       </div>