+++ modules/survey_accounts/test/survey_accountsTest.php
@@ -27,21 +27,15 @@ require_once __DIR__ .
 class Survey_AccountsTestIntegrationTest extends LorisIntegrationTest
 {
     // UI location on the page
-    static $pscid      = "#surveyAccounts_filter > div > div > fieldset >".
-    " div:nth-child(2) > div > div > input";
-    static $visit      = "#surveyAccounts_filter > div > div > fieldset >".
-    " div:nth-child(3) > div > div > select";
-    static $email      = "#surveyAccounts_filter > div > div > fieldset >".
-    " div:nth-child(4) > div > div > input";
-    static $instrument = "#surveyAccounts_filter > div > div > fieldset >".
-    " div:nth-child(5) > div > div > select";
+    static $pscid      = 'input[name="pscid"]';
+    static $visit      = 'select[name="visit"]';
+    static $email      = 'input[name="email"]';
+    static $instrument = 'select[name="instrument"]';
     // clear filter button
-    static $clearFilter = ".col-sm-9 > .btn";
-    static $add         = "#default-panel > div > div > div.table-header >".
-    " div > div > div:nth-child(2) > button:nth-child(1)";
+    static $clearFilter = ".nav-tabs a";
+    static $add         = "div:nth-child(2) > .btn:nth-child(1)";
     // header of the table
-    static $table = "#default-panel > div > div > div.table-header".
-    " > div > div > div:nth-child(1)";
+    static $table = ".table-header > .row > div > div:nth-child(1)";
     /**
      * Insert testing data
      *
@@ -183,7 +177,7 @@ class Survey_AccountsTestIntegrationTest extends LorisIntegrationTest
      */
     function testSurveyAccountsDoespageLoad()
     {
-        $this->setupPermissions(["user_accounts"]);
+        $this->setupPermissions(["survey_accounts_view"]);
         $this->safeGet($this->url . "/survey_accounts/");
         $bodyText
             = $this->webDriver->findElement(WebDriverBy::cssSelector("body"))
@@ -287,7 +281,7 @@ class Survey_AccountsTestIntegrationTest extends LorisIntegrationTest
     function _testFilter($element,$table,$records,$value)
     {
         // get element from the page
-        if (strpos($element, "select") == false) {
+        if (strpos($element, "select") === false) {
             $this->safeFindElement(WebDriverBy::cssSelector($element));
             $this->webDriver->executescript(
                 "input = document.querySelector('$element');